hastyherald/frontend/requirements.txt
websocket-client
Flask